If you spend too much time thinking about a thing, you'll never get it done

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,6 +1,5 @@
 #include "function_pointers.h"
 #include <stdio.h>
-#include <stddef.h>
 
 /**
  * array_iterator - perform function on each element of array
@@ -12,7 +11,8 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned int i;
+
 	if (array == NULL || action == NULL)
 		return;
 	for (i = 0; i < size; i++)
